Split into core and cli crate to prepare for wasm
This commit splits up the library and the binary into separate creates
"core" and "cli" within a shared cargo workspace. This will make it
easier to add a WebAssembly (wasm) target that uses the same backend
(core). Generating both a wasm and a non-wasm binary from a single crete
seems more difficult.

This commit also polishes and document the core crate.