Add support for included png and jpg images
When processing a markdown file, start two operations:
- Convert to pure LaTeX and display result to user. If the markdown
  document contains any images, generate a commented-out
  \includegraphics command
- Start downloading any included images. Once all images are downloaded,
  generate new LaTeX code where the \includegraphics commands are no
  longer commented out and package it together with the images into a
  `.zip` file. Show a button to save the `.zip` file.

Currently, only `.png`, `.jpg`, and `.jpeg` images are supported. It
should be easy to support also `.svg` by converting it to `.png` on the
JavaScript side.

Currently, both tasks run on the main thread. The image files are
downloaded asynchronously, so this is OK. But generating the `.zip` file
can also take some time, and this currently blocks the main thread.